+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 ## Unreleased
 
+#### 🚀 Updates
+
+- Added `Display` to all schema types, so that TypeScript-like type strings can be generated.
+
 #### ⚙️ Internal
 
 - Updated dependencies.